+++ HISTORY.rst
@@ -2,6 +2,27 @@
 History
 =======
 
+unreleased
+----------
+
+Added:
+^^^^^^
+- Added a function to sample times of arrivals of reflections based on a Poisson process (PR #176)
+- Added a function to sample the amplitude or sign of reflections (PR #176)
+- Added functions to calculate the average total number of reflections in a room and their density. (PR #174)
+- All Lundeby related EDC processing in the `pyrato.edc` module now supports frequency dependent smoothing parameters (PR #177)
+- ``center_time`` to compute the room-acoustic center time (:math:`T_s`) from an energy decay curve (PR #170).
+
+Deprecated:
+^^^^^^^^^^^
+- The `freq` parameter in the `pyrato.edc` module was renamed to `smoothing_parameter` and will be deprecated in pyrato v1.3.0
+
+Changed:
+^^^^^^^^
+- Updated minimum required dependency version to scipy>=1.11.0 (PR #170)
+- Constrain Sphinx to versions >=v8.0.0 and <v9.0.0 (PR #168)
+
+
 1.0.1 (2026-05-05)
 ------------------